+++ src/data.ts
@@ -1,8 +1,0 @@
-/**
- * Copyright © 2023-2026 Blockchain Commons, LLC
- * Copyright © 2025-2026 Parity Technologies
- *
- *
- * An idiom for a block of data used throughout LifeHash.
- */
-export type Data = Uint8Array;