+++ src/sha256.ts
@@ -1,8 +1,0 @@
-/**
- * Copyright © 2023-2026 Blockchain Commons, LLC
- * Copyright © 2025-2026 Parity Technologies
- *
- *
- * Re-export SHA-256 from @blockchaincommons/crypto for consistency.
- */
-export { sha256 } from "@blockchaincommons/crypto";